requirements.txt
Flask
flask_wtf
grab
pycurl
requests
lxml
xmltodict
pymongo
mysqlclient
python-dotenv
